Refactoring of Automate Engine code
Separated State Machine specific code into a separate module
Moved the state machine spec into engine_validation directory

https://bugzilla.redhat.com/show_bug.cgi?id=1215990
